Rename directories for views and controllers
