+++ src/main/java/com/thebois/controllers/game/README.md
@@ -1,0 +1,5 @@
+# com.thebois.controllers.game
+
+## UML Class Diagram
+
+![com.thebois.controllers.game](../../../../../../../documents/diagrams/com.thebois.controllers.game.jpg "com.thebois.controllers.game")
