--- src/main/java/com/thebois/controllers/gamecontrollers/README.md
@@ -1,5 +1,0 @@
-# com.thebois.controllers.gamecontrollers
-
-## UML Class Diagram
-
-![com.thebois.controllers.gamecontrollers](./../../../../../../../documents/diagrams/com.thebois.controllers.gamecontrollers.jpg "com.thebois.controllers.gamecontrollers")
