+++ src/main/java/com/thebois/controllers/info/README.md
@@ -1,0 +1,5 @@
+# com.thebois.controllers.info
+
+## UML Class Diagram
+
+![com.thebois.controllers.info](../../../../../../../documents/diagrams/com.thebois.controllers.info.jpg "com.thebois.controllers.info")
