+++ src/main/java/com/thebois/controllers/info/RoleController.java
@@ -1,4 +1,4 @@
-package com.thebois.controllers.infocontrollers;
+package com.thebois.controllers.info;
 
 import java.util.AbstractMap;
 import java.util.stream.Collectors;
@@ -11,9 +11,9 @@ import com.thebois.models.beings.roles.AbstractRole;
 import com.thebois.models.beings.roles.IRoleAllocator;
 import com.thebois.models.beings.roles.RoleFactory;
 import com.thebois.models.beings.roles.RoleType;
-import com.thebois.views.infoviews.IActorView;
-import com.thebois.views.infoviews.RoleView;
-import com.thebois.views.infoviews.SpinnerButton;
+import com.thebois.views.info.IActorView;
+import com.thebois.views.info.RoleView;
+import com.thebois.views.info.SpinnerButton;
 
 /**
  * Controls the interactions from the player regarding role allocations.
