+++ src/main/java/com/thebois/controllers/info/package-info.java
@@ -2,4 +2,4 @@
  * Controllers that are responsible for info panels.
  */
 
-package com.thebois.controllers.infocontrollers;
+package com.thebois.controllers.info;
