--- src/main/java/com/thebois/controllers/infocontrollers/README.md
@@ -1,5 +1,0 @@
-# com.thebois.controllers.infocontrollers
-
-## UML Class Diagram
-
-![com.thebois.controllers.infocontrollers](../../../../../../../documents/diagrams/com.thebois.controllers.infocontrollers.jpg "com.thebois.controllers.infocontrollers")
