+++ src/main/java/com/thebois/views/debug/BeingPathDebugView.java
@@ -7,7 +7,7 @@ import com.badlogic.gdx.graphics.g2d.Batch;
 import com.thebois.models.Position;
 import com.thebois.models.beings.IBeing;
 import com.thebois.models.beings.IBeingGroup;
-import com.thebois.views.gameviews.IView;
+import com.thebois.views.game.IView;
 import com.thebois.views.TextureUtils;
 
 /**
