+++ src/main/java/com/thebois/views/game/ColonyView.java
@@ -1,4 +1,4 @@
-package com.thebois.views.gameviews;
+package com.thebois.views.game;
 
 import java.util.HashMap;
 import java.util.Map;
