+++ src/main/java/com/thebois/views/game/README.md
@@ -1,0 +1,5 @@
+# com.thebois.views.game
+
+## UML Class Diagram
+
+![com.thebois.views.game](../../../../../../../documents/diagrams/com.thebois.views.game.jpg "com.thebois.views.game")
