+++ src/main/java/com/thebois/views/game/package-info.java
@@ -2,4 +2,4 @@
  * Views that handle rendering of gameplay.
  */
 
-package com.thebois.views.gameviews;
+package com.thebois.views.game;
