--- src/main/java/com/thebois/views/gameviews/README.md
@@ -1,5 +1,0 @@
-# com.thebois.views.gameviews
-
-## UML Class Diagram
-
-![com.thebois.views.gameviews](./../../../../../../../documents/diagrams/com.thebois.views.gameviews.jpg "com.thebois.views.gameviews")
