+++ src/main/java/com/thebois/views/info/README.md
@@ -1,0 +1,5 @@
+# com.thebois.views.info
+
+## UML Class Diagram
+
+![com.thebois.views.info](../../../../../../../documents/diagrams/com.thebois.views.info.jpg "com.thebois.views.info")
