+++ src/main/java/com/thebois/views/info/RoleView.java
@@ -1,4 +1,4 @@
-package com.thebois.views.infoviews;
+package com.thebois.views.info;
 
 import java.util.AbstractMap;
 import java.util.ArrayList;
