+++ src/main/java/com/thebois/views/info/SpinnerButton.java
@@ -1,4 +1,4 @@
-package com.thebois.views.infoviews;
+package com.thebois.views.info;
 
 import java.util.ArrayList;
 import java.util.Collection;
