+++ src/main/java/com/thebois/views/info/package-info.java
@@ -2,4 +2,4 @@
  * Views that that present information and not gameplay.
  */
 
-package com.thebois.views.infoviews;
+package com.thebois.views.info;
