--- src/main/java/com/thebois/views/infoviews/README.md
@@ -1,5 +1,0 @@
-# com.thebois.views.infoviews
-
-## UML Class Diagram
-
-![com.thebois.views.infoviews](./../../../../../../../documents/diagrams/com.thebois.views.infoviews.jpg "com.thebois.views.infoviews")
